final reward should be set by env, not agent

+++ src/rl/agent/human_player.rs
@@ -46,7 +46,7 @@ impl Agent for HumanPlayer {
         }
     }
 
-    fn finish_round(&mut self, _single_res: i32, _final_state: Array2<f32>) {}
+    fn finish_round(&mut self, _single_res: f32, _final_state: Array2<f32>) {}
 
     fn get_learning_rate(&self) -> f32 {
         42.
